Overload authorize directive to check against service context
This enables checking permissions when a service request is
constructed synthetically.

An example where this may be useful is in the case where an
authorization token is not available as a header but can be obtained
elsewhere, such as a session cookie. The service context may then be
explicitly constructed with said token and reuse the existing
permissions checking functionality.