Move database-related functionality to separate project
This committ includes a breaking change. The database-specific utility
"Converters" trait threw an exception "DatabaseException" defined in
the rest package, thus breaking the dependency graph. The solution was
to move the DatabaseException class from rest to database and not
inherit ServiceExceptio any more.

Unfortunately, the rest classes also require the database exception in
propagating errors so this funtionality has been removed. The
rationale is:

  1. Database exceptions are rare and result in 500 errors anyway
  making the status code opaque to what actual error caused it.

  2. In core 2.0, an improved tracing framework will make diagnosing
  and following database errors easier, thereby attenuating the need
  to forward details on service exceptions in responses.